requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
Pillow==8.3.2
PyPDF2==1.26.0
python-docx==0.8.11
pandas==1.3.3
openpyxl==3.0.9
matplotlib==3.4.3
seaborn==0.11.2
numpy==1.21.2